employee-api/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pydantic
python-dotenv
pytest
requests